+++ api/operator/v1beta1/vmcluster_types.go
@@ -494,8 +494,7 @@ type VMBackup struct {
 	// AcceptEULA accepts enterprise feature usage, must be set to true.
 	// otherwise backupmanager cannot be added to single/cluster version.
 	// https://victoriametrics.com/legal/esa/
-	// Deprecated: use license.key or license.keyRef instead
-	// +deprecated
+	// +deprecated={deprecated_in: "v0.61.0", removed_in: "v0.69.0", replacements: {license.key,license.keyRef}}
 	// +optional
 	AcceptEULA bool `json:"acceptEULA"`
 	// SnapshotCreateURL overwrites url for snapshot create